requirements.txt
requests>=2.31.0
mysql-connector-python>=8.1.0
python-dotenv>=1.0.0
selenium>=4.10.0
webdriver-manager>=4.0.0
python-dateutil>=2.8.2